move src/containers/settings/ into appropriate sub-folder(s) of src/components/settings/
